backend/requirements.txt
alembic==1.15.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
flask-mysql-connector==1.1.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
git-filter-repo==2.47.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mysql-connector-python==9.2.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
PyMySQL==1.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
tk==0.1.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
